+++ core/src/test/java/io/questdb/cairo/TableWriterTest.java
@@ -315,7 +315,7 @@ public class TableWriterTest extends AbstractCairoTest {
         String abcColumnNamePattern = Files.SEPARATOR + "abc.d";
         testAddColumnRecoverableFault(new FilesFacadeImpl() {
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public int openRW(LPSZ name, long opts) {
                 if (Chars.contains(name, abcColumnNamePattern)) {
                     return -1;
                 }
