+++ core/src/test/java/io/questdb/griffin/engine/functions/catalogue/WalTableListFunctionFactoryTest.java
@@ -42,7 +42,7 @@ public class WalTableListFunctionFactoryTest extends AbstractGriffinTest {
             private int attempt = 0;
 
             @Override
-            public long openRW(LPSZ name, long opts) {
+            public int openRW(LPSZ name, long opts) {
                 if (Chars.contains(name, "x.d.1") && attempt++ == 0) {
                     return -1;
                 }
